+++ +nix/SampledDimension.m
@@ -28,7 +28,7 @@ classdef SampledDimension < nix.Entity
                 index = index - 1;
             end
             func_name = strcat(obj.alias, '::position_at');
-            posAt = nix_mx(func_name, obj.nix_handle, uint64(index));
+            posAt = nix_mx(func_name, obj.nix_handle, index);
         end
 
         function axis = axis(obj, count, startIndex)
@@ -39,8 +39,7 @@ classdef SampledDimension < nix.Entity
                 startIndex = startIndex - 1;
             end
             func_name = strcat(obj.alias, '::axis');
-            axis = nix_mx(func_name, obj.nix_handle, uint64(count), uint64(startIndex));
+            axis = nix_mx(func_name, obj.nix_handle, count, startIndex);
         end
     end
 end
-
